removed action reducer and store and added loginAction
