+++ src/index.js
@@ -3,16 +3,14 @@ import ReactDOM from "react-dom";
 import App from "./components/App";
 import * as serviceWorker from "./serviceWorker";
 import { BrowserRouter } from "react-router-dom";
-import { Provider } from 'react-redux';
-import { store, history } from './components/store'
 import "./index.css";
 
 ReactDOM.render(
   
   <BrowserRouter>
-  <Provider store={store}>
+  
     <App />
-    </Provider>
+    
   </BrowserRouter>
   ,
   document.getElementById("root")
